// src/lib/api/generated/model/createUserDtoRole.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

/**
 * Defaults to CLIENT when omitted.
 */
export type CreateUserDtoRole = typeof CreateUserDtoRole[keyof typeof CreateUserDtoRole];


export const CreateUserDtoRole = {
  ADMIN: 'ADMIN',
  CLIENT: 'CLIENT',
} as const;